Move declarations into separate module
Ref. #47, #382
